backend/requirements.txt
flask==3.0.0
flask-cors==4.0.0
psycopg2-binary==2.9.9
playwright==1.41.0
python-dotenv==1.0.0
gunicorn==21.2.0
